# src/main/java/com/example/BellTestProject/model/User.java
package com.example.BellTestProject.model;

import javax.persistence.*;

@Entity
@Table(name = "users")
public class User {

    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    @Column(name = "user_id", unique = true, nullable = false)
    private int id;

    @Column(name = "first_name" , length = 45 , nullable = false)
    private String firstName;

    @Column(name = "position" , length = 45 , nullable = false)
    private String position;

    @Column(name = "phone")
    private int phone;

    @Column(name = "is_identified")
    private boolean isIdentified;

   @ManyToOne(cascade = CascadeType.MERGE)
   @JoinColumn(name = "office_id")
   private Office office;

   public User() {

   }

    public int getId() {
        return id;
    }

    public void setId(int id) {
        this.id = id;
    }

    public String getFirstName() {
        return firstName;
    }

    public void setFirstName(String firstName) {
        this.firstName = firstName;
    }

    public String getPosition() {
        return position;
    }

    public void setPosition(String position) {
        this.position = position;
    }

    public int getPhone() {
        return phone;
    }

    public void setPhone(int phone) {
        this.phone = phone;
    }

    public boolean isIdentified() {
        return isIdentified;
    }

    public void setIdentified(boolean identified) {
        isIdentified = identified;
    }

    public Office getOffice() {
        return office;
    }

    public void setOffice(Office office) {
        this.office = office;
    }
}